Add --no-nix method to installation instructions (#46)
* Add --no-nix method to installation instructions

Please let me know if this is the correct. It worked on my `antergos` setup. The reason for me to post this is that I could not get `nix` working on my system. Upon asking @ChrisPenner on `gitter` I got the `--no-nix` clue. The editor starts up. Although I still don't know how to use it.

* Minor Corrections